spline #108 refactoring step 4: Extract write plugins
