--- core/src/main/scala/za/co/absa/spline/harvester/builder/SourceUri.scala
@@ -1,41 +1,0 @@
-/*
- * Copyright 2019 ABSA Group Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package za.co.absa.spline.harvester.builder
-
-import org.apache.spark.sql.SparkSession
-import org.apache.spark.sql.catalyst.TableIdentifier
-
-object SourceUri {
-  def forKafka(topic: String): String = s"kafka:$topic"
-
-  def forJDBC(connectionUrl: String, table: String): String = s"$connectionUrl:$table"
-
-  def forCassandra(keyspace: String, table: String): String = s"cassandra:$keyspace:$table"
-
-  def forMongoDB(connectionUrl: String, database: String, collection: String): String = s"$connectionUrl/$database.$collection"
-
-  def forElastiSearch(server: String, indexDocType: String): String = s"elasticsearch://$server/$indexDocType"
-
-  def forTable(tableIdentifier: TableIdentifier)
-    (session: SparkSession): String = {
-    val catalog = session.catalog
-    val TableIdentifier(tableName, maybeTableDatabase) = tableIdentifier
-    val databaseName = maybeTableDatabase getOrElse catalog.currentDatabase
-    val databaseLocation = catalog.getDatabase(databaseName).locationUri.stripSuffix("/")
-    s"$databaseLocation/${tableName.toLowerCase}"
-  }
-}
