+++ core/src/main/scala/za/co/absa/spline/harvester/plugin/impl/CobrixPlugin.scala
@@ -20,7 +20,7 @@ import org.apache.spark.sql.execution.datasources.LogicalRelation
 import org.apache.spark.sql.sources.BaseRelation
 import za.co.absa.commons.reflect.ReflectionUtils.extractFieldValue
 import za.co.absa.commons.reflect.extractors.SafeTypeMatchingExtractor
-import za.co.absa.spline.harvester.builder.SourceIdentifier
+import za.co.absa.spline.harvester.builder.{SourceId, SourceIdentifier}
 import za.co.absa.spline.harvester.plugin.Plugin.Params
 import za.co.absa.spline.harvester.plugin.impl.CobrixPlugin.`_: CobolRelation`
 import za.co.absa.spline.harvester.plugin.{BaseRelationPlugin, Plugin}
@@ -32,7 +32,7 @@ class CobrixPlugin extends Plugin with BaseRelationPlugin {
 
     case (`_: CobolRelation`(cr), _) =>
       val sourceDir = extractFieldValue[String](cr, "sourceDir")
-      (SourceIdentifier.forCobol(sourceDir), Map.empty)
+      (SourceId.forCobol(sourceDir), Map.empty)
   }
 }
