+++ core/src/main/scala/za/co/absa/spline/harvester/plugin/impl/ElasticSearchPlugin.scala
@@ -16,32 +16,47 @@
 
 package za.co.absa.spline.harvester.plugin.impl
 
-import org.apache.spark.sql.execution.datasources.LogicalRelation
+import org.apache.spark.sql.SaveMode
+import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
+import org.apache.spark.sql.execution.datasources.{LogicalRelation, SaveIntoDataSourceCommand}
 import org.apache.spark.sql.sources.BaseRelation
 import org.elasticsearch.spark.cfg.SparkSettings
 import za.co.absa.commons.reflect.ReflectionUtils.extractFieldValue
 import za.co.absa.commons.reflect.extractors.SafeTypeMatchingExtractor
-import za.co.absa.spline.harvester.builder.SourceIdentifier
+import za.co.absa.spline.harvester.builder.{SourceId, SourceIdentifier}
 import za.co.absa.spline.harvester.plugin.Plugin.Params
-import za.co.absa.spline.harvester.plugin.impl.ElasticSearchPlugin.`_: ElasticsearchRelation`
-import za.co.absa.spline.harvester.plugin.{BaseRelationPlugin, Plugin}
+import za.co.absa.spline.harvester.plugin.impl.ElasticSearchPlugin._
+import za.co.absa.spline.harvester.plugin.{BaseRelationPlugin, DataSourceTypePlugin, Plugin}
 
 
-class ElasticSearchPlugin extends Plugin with BaseRelationPlugin {
+class ElasticSearchPlugin
+  extends Plugin
+    with BaseRelationPlugin
+    with DataSourceTypePlugin {
+
+  import za.co.absa.commons.ExtractorImplicits._
 
   override def baseRelProcessor: PartialFunction[(BaseRelation, LogicalRelation), (SourceIdentifier, Params)] = {
-
     case (`_: ElasticsearchRelation`(esr), _) =>
       val parameters = extractFieldValue[SparkSettings](esr, "cfg")
       val server = parameters.getProperty("es.nodes")
       val indexDocType = parameters.getProperty("es.resource")
-      (SourceIdentifier.forElasticSearch(server, indexDocType), Map.empty)
+      (SourceId.forElasticSearch(server, indexDocType), Map.empty)
+  }
+
+  override def dataSourceTypeProcessor: PartialFunction[(AnyRef, SaveIntoDataSourceCommand), (SourceIdentifier, SaveMode, LogicalPlan, Params)] = {
+    case (st, cmd) if st == "es" || ElasticSearchSourceExtractor.matches(st) =>
+      val indexDocType = cmd.options("path")
+      val server = cmd.options("es.nodes")
+      (SourceId.forElasticSearch(server, indexDocType), cmd.mode, cmd.query, cmd.options)
   }
 }
 
 object ElasticSearchPlugin {
 
   object `_: ElasticsearchRelation` extends SafeTypeMatchingExtractor[AnyRef]("org.elasticsearch.spark.sql.ElasticsearchRelation")
+
+  private object ElasticSearchSourceExtractor extends SafeTypeMatchingExtractor(classOf[org.elasticsearch.spark.sql.DefaultSource15])
 
 }
