+++ core/src/main/scala/za/co/absa/spline/harvester/plugin/impl/ExcelPlugin.scala
@@ -23,7 +23,7 @@ import org.apache.spark.sql.execution.datasources.LogicalRelation
 import org.apache.spark.sql.sources.BaseRelation
 import za.co.absa.commons.reflect.ReflectionUtils.extractFieldValue
 import za.co.absa.commons.reflect.extractors.SafeTypeMatchingExtractor
-import za.co.absa.spline.harvester.builder.SourceIdentifier
+import za.co.absa.spline.harvester.builder.{SourceId, SourceIdentifier}
 import za.co.absa.spline.harvester.plugin.Plugin.Params
 import za.co.absa.spline.harvester.plugin.impl.ExcelPlugin._
 import za.co.absa.spline.harvester.plugin.{BaseRelationPlugin, Plugin}
@@ -41,7 +41,7 @@ class ExcelPlugin(pathQualifier: PathQualifier) extends Plugin with BaseRelation
       val inputStream = extractExcelInputStream(excelRelation.workbookReader)
       val path = extractFieldValue[org.apache.hadoop.fs.Path](inputStream, "file")
       val qualifiedPath = pathQualifier.qualify(path.toString)
-      (SourceIdentifier.forExcel(qualifiedPath),
+      (SourceId.forExcel(qualifiedPath),
         extractExcelParams(excelRelation) + ("header" -> excelRelation.header.toString))
   }
 }
