+++ core/src/main/scala/za/co/absa/spline/harvester/plugin/impl/JDBCPlugin.scala
@@ -16,28 +16,38 @@
 
 package za.co.absa.spline.harvester.plugin.impl
 
-import org.apache.spark.sql.execution.datasources.LogicalRelation
-import org.apache.spark.sql.execution.datasources.jdbc.JDBCOptions
+import org.apache.spark.sql.SaveMode
+import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
+import org.apache.spark.sql.execution.datasources.jdbc.{JDBCOptions, JdbcRelationProvider}
+import org.apache.spark.sql.execution.datasources.{LogicalRelation, SaveIntoDataSourceCommand}
 import org.apache.spark.sql.sources.BaseRelation
 import za.co.absa.commons.reflect.ReflectionUtils.extractFieldValue
 import za.co.absa.commons.reflect.extractors.{AccessorMethodValueExtractor, SafeTypeMatchingExtractor}
-import za.co.absa.spline.harvester.builder.SourceIdentifier
+import za.co.absa.spline.harvester.builder.{SourceId, SourceIdentifier}
 import za.co.absa.spline.harvester.plugin.Plugin.Params
 import za.co.absa.spline.harvester.plugin.impl.JDBCPlugin._
-import za.co.absa.spline.harvester.plugin.{BaseRelationPlugin, Plugin}
+import za.co.absa.spline.harvester.plugin.{BaseRelationPlugin, DataSourceTypePlugin, Plugin}
 
 
-class JDBCPlugin extends Plugin with BaseRelationPlugin {
-
+class JDBCPlugin
+  extends Plugin
+    with BaseRelationPlugin
+    with DataSourceTypePlugin {
 
   override def baseRelProcessor: PartialFunction[(BaseRelation, LogicalRelation), (SourceIdentifier, Params)] = {
-
     case (`_: JDBCRelation`(jr), _) =>
       val jdbcOptions = extractFieldValue[JDBCOptions](jr, "jdbcOptions")
       val url = extractFieldValue[String](jdbcOptions, "url")
       val params = extractFieldValue[Map[String, String]](jdbcOptions, "parameters")
       val TableOrQueryFromJDBCOptionsExtractor(toq) = jdbcOptions
-      (SourceIdentifier.forJDBC(url, toq), params)
+      (SourceId.forJDBC(url, toq), params)
+  }
+
+  override def dataSourceTypeProcessor: PartialFunction[(AnyRef, SaveIntoDataSourceCommand), (SourceIdentifier, SaveMode, LogicalPlan, Params)] = {
+    case (st, cmd) if st == "jdbc" || st.isInstanceOf[JdbcRelationProvider] =>
+      val jdbcConnectionString = cmd.options("url")
+      val tableName = cmd.options("dbtable")
+      (SourceId.forJDBC(jdbcConnectionString, tableName), cmd.mode, cmd.query, Map.empty)
   }
 }
